refactor: move `json` package to `pkg/x/json`
